# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KOSMOS API",
    "description": "KOSMOS AI Governance Platform - Multi-agent orchestration with Pentarchy voting.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/votes/proposals": {
      "post": {
        "tags": [
          "governance"
        ],
        "summary": "Create Proposal",
        "description": "Create a new proposal and initiate Pentarchy voting.",
        "operationId": "create_proposal_api_v1_votes_proposals_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "governance"
        ],
        "summary": "List Proposals",
        "description": "List all proposals with optional filtering.",
        "operationId": "list_proposals_api_v1_votes_proposals_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(pending|approved|rejected|escalated)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProposalSummary"
                  },
                  "title": "Response List Proposals Api V1 Votes Proposals Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/votes/proposals/{proposal_id}": {
      "get": {
        "tags": [
          "governance"
        ],
        "summary": "Get Proposal",
        "description": "Get details of a specific proposal.",
        "operationId": "get_proposal_api_v1_votes_proposals__proposal_id__get",
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Proposal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/votes/proposals/{proposal_id}/vote": {
      "post": {
        "tags": [
          "governance"
        ],
        "summary": "Manual Vote",
        "description": "Manually add a vote (for testing or override).",
        "operationId": "manual_vote_api_v1_votes_proposals__proposal_id__vote_post",
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Proposal Id"
            }
          },
          {
            "name": "agent",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent"
            }
          },
          {
            "name": "vote",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vote"
            }
          },
          {
            "name": "score",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "Score"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "title": "Reasoning"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/votes/proposals/{proposal_id}/resolve": {
      "post": {
        "tags": [
          "governance"
        ],
        "summary": "Resolve Proposal",
        "description": "Force resolution of a pending proposal.",
        "operationId": "resolve_proposal_api_v1_votes_proposals__proposal_id__resolve_post",
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Proposal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProposalRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 10,
            "title": "Description"
          },
          "cost": {
            "type": "number",
            "minimum": 0.0,
            "title": "Cost",
            "default": 0.0
          },
          "risk_level": {
            "type": "string",
            "pattern": "^(low|medium|high|critical)$",
            "title": "Risk Level",
            "default": "medium"
          },
          "context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "auto_execute": {
            "type": "boolean",
            "title": "Auto Execute",
            "description": "Auto-execute if approved",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "title",
          "description"
        ],
        "title": "ProposalRequest",
        "description": "Request to create a proposal for Pentarchy voting."
      },
      "ProposalResponse": {
        "properties": {
          "proposal_id": {
            "type": "string",
            "title": "Proposal Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "cost": {
            "type": "number",
            "title": "Cost"
          },
          "risk_level": {
            "type": "string",
            "title": "Risk Level"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "votes": {
            "items": {
              "$ref": "#/components/schemas/VoteResult"
            },
            "type": "array",
            "title": "Votes"
          },
          "final_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Score"
          },
          "threshold": {
            "type": "number",
            "title": "Threshold"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          }
        },
        "type": "object",
        "required": [
          "proposal_id",
          "title",
          "description",
          "cost",
          "risk_level",
          "status",
          "votes",
          "final_score",
          "threshold",
          "created_at",
          "resolved_at"
        ],
        "title": "ProposalResponse",
        "description": "Response with proposal details and voting results."
      },
      "ProposalSummary": {
        "properties": {
          "proposal_id": {
            "type": "string",
            "title": "Proposal Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "final_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Score"
          },
          "vote_count": {
            "type": "integer",
            "title": "Vote Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "proposal_id",
          "title",
          "status",
          "final_score",
          "vote_count",
          "created_at"
        ],
        "title": "ProposalSummary",
        "description": "Summary of a proposal."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VoteResult": {
        "properties": {
          "agent": {
            "type": "string",
            "title": "Agent"
          },
          "vote": {
            "type": "string",
            "title": "Vote"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "reasoning": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasoning"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "agent",
          "vote",
          "score",
          "reasoning",
          "timestamp"
        ],
        "title": "VoteResult",
        "description": "Individual agent vote."
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}